refactor: use external mathdroid-covid19 plugin
