--- plugins/gatsby-source-covid19/.gitignore
@@ -1,7 +1,0 @@
-/__tests__
-/yarn.lock
-/*.js
-/*.js.map
-/*.d.*
-!index.js
-built
