--- plugins/gatsby-source-covid19/index.js
@@ -1,1 +1,0 @@
-//boop
