--- plugins/gatsby-source-covid19/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es5",
-    "lib": ["dom", "es6"],
-    "sourceMap": true,
-    "jsx": "react",
-    "noUnusedLocals": true,
-    "forceConsistentCasingInFileNames": true,
-    "noImplicitThis": true,
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "declaration": true,
-    "declarationMap": true,
-    "allowSyntheticDefaultImports": true,
-    "strict": true,
-    "typeRoots": ["./node_modules/@types"],
-    "types": ["node"],
-    "esModuleInterop": true,
-    "noEmit": false,
-    "noEmitOnError": true,
-    "outDir": ".",
-    "skipLibCheck": true
-  },
-  "include": ["./src/**/*"],
-  "exclude": ["./node_modules/*"]
-}
